pyproject.toml
[project]
name = "network-firewall"
version = "0.1.0"
description = "简易网络防火墙，基于PyDivert和PyQt6实现"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphviz>=0.20.3",
    "mss>=10.0.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pydivert>=2.1.0",
    "pyqt6>=6.8.1",
    "pytest>=8.3.5",
    "pytest-qt>=4.4.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
firewall = "main:main"

[tool.setuptools]
packages = ["firewall", "firewall.core", "firewall.ui", "firewall.resources"]
